Move: 인증 가드 파일 이동
